Add SWRSokuRoll module
This code was originally created by Ichirin, then modified by delthas to
integrate it into this repository, replacing the MemoryModule files with
a submodule, simplifying the INI logic, and waiting for the window to be
shown to inject the module.

SokuRoll.inl is made from SokuRoll.dll, which was made by FireSeal. This
mod merely packages this library and injects it into the game.