Update WindowResizer.
Fixed issue with sizing.
Game rendering area was slightly smaller than the usual resolution.
Readding old styles fixed the issue.

Fixed(?) issue with the game changing focus to the previous window.
Seems to be caused by the black bars taking focus and not giving it back
to the old window. Fixed by setting focus back manually, still has some
issues but mostly fixes it.

Added a toggle for always-on-top/sticky mode on alt+p by default,
configurable in the ini.

Added toggles for integer scaling on alt+1/2/3/4/5/6 for 1x/2x/3x/etc...
Should be changed to proper presets maybe but this is fine for now.